Add a new test class
