+++ pulsar-broker/src/test/java/org/apache/pulsar/client/impl/MessageChunkingSharedTest.java
@@ -21,11 +21,9 @@ package org.apache.pulsar.client.impl;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertNotNull;
-import static org.testng.Assert.assertNull;
 import static org.testng.Assert.assertTrue;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
-import java.lang.reflect.Field;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -214,123 +212,13 @@ public class MessageChunkingSharedTest extends ProducerConsumerBase {
         return Schema.STRING.decode(payload);
     }
 
-    @Test
-    public void testSendChunkMessageWithSameSequenceID() throws Exception {
-        this.conf.setBrokerDeduplicationEnabled(true);
-        restartBroker();
-        String topicName = "persistent://my-property/my-ns/testSendChunkMessageWithSameSequenceID";
-        String producerName = "test-producer";
-        @Cleanup
-        Consumer<String> consumer = pulsarClient
-                .newConsumer(Schema.STRING)
-                .subscriptionName("test-sub")
-                .topic(topicName)
-                .subscribe();
-        @Cleanup
-        Producer<String> producer = pulsarClient
-                .newProducer(Schema.STRING)
-                .producerName(producerName)
-                .topic(topicName)
-                .enableChunking(true)
-                .enableBatching(false)
-                .create();
-        int messageSize = 6000; // payload size in KB
-        String message = "a".repeat(messageSize * 1000);
-        producer.newMessage().value(message).sequenceId(10).send();
-        Message<String> msg = consumer.receive(10, TimeUnit.SECONDS);
-        assertNotNull(msg);
-        assertTrue(msg.getMessageId() instanceof ChunkMessageIdImpl);
-        assertEquals(msg.getValue(), message);
-        producer.newMessage().value(message).sequenceId(10).send();
-        msg = consumer.receive(3, TimeUnit.SECONDS);
-        assertNull(msg);
-    }
-
-    @Test
-    public void testDuplicateForChunkMessage() throws Exception {
-        this.conf.setBrokerDeduplicationEnabled(true);
-        restartBroker();
-        String topicName = "persistent://my-property/my-ns/testDuplicateForChunkMessage";
-        String producerName = "test-producer";
-        @Cleanup
-        Consumer<String> consumer = pulsarClient
-                .newConsumer(Schema.STRING)
-                .subscriptionName("test-sub")
-                .topic(topicName)
-                .subscribe();
-        @Cleanup
-        Producer<String> partProducer = pulsarClient
-                .newProducer(Schema.STRING)
-                .producerName(producerName)
-                .topic(topicName)
-                .enableChunking(true)
-                .enableBatching(false)
-                .create();
-        int messageSize = 6000; // payload size in KB
-        String message = "a".repeat(messageSize * 1000);
-        partProducer.newMessage().value(message).send();
-        Message<String> msg = consumer.receive(5, TimeUnit.SECONDS);
-        assertNotNull(msg);
-        assertTrue(msg.getMessageId() instanceof ChunkMessageIdImpl);
-        assertEquals(msg.getValue(), message);
-
-        Field msgIdGenerator = ProducerImpl.class.getDeclaredField("msgIdGenerator");
-        msgIdGenerator.setAccessible(true);
-        assertEquals(msg.getSequenceId() + 1, msgIdGenerator.get(partProducer));
-
-        String message2 = "b".repeat(messageSize * 2);
-        partProducer.newMessage().value(message2).send();
-        Message<String> msg2 = consumer.receive(5, TimeUnit.SECONDS);
-        assertFalse(msg2.getMessageId() instanceof ChunkMessageIdImpl);
-        assertEquals(msg2.getValue(), message2);
-
-        long sequenceID = (long) msgIdGenerator.get(partProducer) + 1024L;
-        String message3 = "c".repeat(messageSize * 1000);
-        partProducer.newMessage().value(message3).sequenceId(sequenceID).send();
-        Message<String> msg3 = consumer.receive(5, TimeUnit.SECONDS);
-        assertNotNull(msg3);
-        assertTrue(msg3.getMessageId() instanceof ChunkMessageIdImpl);
-        assertEquals(msg3.getValue(), message3);
-        assertEquals(msg3.getSequenceId(), sequenceID);
-    }
-
-    @Test
-    public void testDeduplicateChunksInSingleChunkMessages() throws Exception {
-        this.conf.setBrokerDeduplicationEnabled(true);
-        restartBroker();
-        String topicName = "persistent://my-property/my-ns/testDeduplicateChunksInSingleChunkMessage";
-        String producerName = "test-producer";
-        @Cleanup
-        Consumer<String> consumer = pulsarClient
-                .newConsumer(Schema.STRING)
-                .subscriptionName("test-sub")
-                .topic(topicName)
-                .subscribe();
-        final PersistentTopic persistentTopic = (PersistentTopic) pulsar.getBrokerService()
-                .getTopicIfExists(topicName).get().orElse(null);
-        assertNotNull(persistentTopic);
-        sendChunk(persistentTopic, producerName, 1, 0, 2);
-        sendChunk(persistentTopic, producerName, 1, 1, 2);
-        sendChunk(persistentTopic, producerName, 1, 1, 2);
-
-        Message<String> message = consumer.receive(15, TimeUnit.SECONDS);
-        assertEquals(message.getData().length, 2);
-
-        sendChunk(persistentTopic, producerName, 2, 0, 3);
-        sendChunk(persistentTopic, producerName, 2, 1, 3);
-        sendChunk(persistentTopic, producerName, 2, 1, 3);
-        sendChunk(persistentTopic, producerName, 2, 2, 3);
-        message = consumer.receive(20, TimeUnit.SECONDS);
-        assertEquals(message.getData().length, 3);
-    }
-
     private static void sendNonChunk(final PersistentTopic persistentTopic,
                                      final String producerName,
                                      final long sequenceId) {
         sendChunk(persistentTopic, producerName, sequenceId, null, null);
     }
 
-    private static void sendChunk(final PersistentTopic persistentTopic,
+    protected static void sendChunk(final PersistentTopic persistentTopic,
                                   final String producerName,
                                   final long sequenceId,
                                   final Integer chunkId,
